moved GalSimDetector to its own source code file
