+++ python/lsst/sims/GalSimInterface/__init__.py
@@ -1,3 +1,4 @@
+from .galSimDetector import *
 from .galSimCelestialObject import *
 from .galSimNoiseAndBackground import *
 from .galSimPSF import *
